Fix cwd error
This error was hapenning once in a while when a change received all approval
necessary to be submitted upstream. This happens because another thread or
process deleted the directory the process started to run, so os.getcwd() will
fail.

We actually don't need to keep stated of the CWD as this is running all gw
commands with full path and even if don't, the CWD would be deleted anyway so,
there would no way to return to it.